Optimize Coord because using of constructor to tranform Coord too vector was taking a lot of time.
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@10 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/layout/SpringElectrical.cpp
@@ -93,7 +93,7 @@ bool SpringElectrical::run() {
 
   k = 5.0;
   k2 = 2.0*k;
-  double maxforce = log(superGraph->numberOfNodes());
+  double maxforce = log(double(superGraph->numberOfNodes()));
   double deltaForce = maxforce / iterations;
 
   for (int count = 0; count < iterations; ++count) {
